refactor: improve AVX, make detection a little nicer
split out isa into four separate files, rather than one long snakepit of #ifdef
pass flags properly to caller, and have the caller set the compile flags and define
set the define as a fixed name with a value 0, 1, 2 or 512, instead of four different defines
