+++ library/CMakeLists.txt
@@ -24,13 +24,13 @@ add_compile_definitions(DPP_OS=${CMAKE_SYSTEM_NAME})
 
 if(${AVX_TYPE} STREQUAL "OFF")
 	include("${CMAKE_CURRENT_SOURCE_DIR}/../cmake/DetectArchitecture.cmake")
-	add_compile_options("${AVX_FLAG}")
-	add_compile_definitions("${AVX_NAME}")
-	message("--- AVX type: ${AVX_NAME}")
+	message("--- AVX type: ${AVX_TYPE}")
 else()
 	message("-- AVX type overridden by configuration: ${AVX_TYPE}")
-	add_compile_definitions("${AVX_TYPE}")
 endif()
+STRING(REPLACE "AVX" "" AVX_TYPE ${AVX_TYPE})
+add_compile_definitions(AVX_TYPE=${AVX_TYPE})
+add_compile_options(${AVX_FLAG})
 
 if(WIN32 AND NOT MINGW)
 	if (NOT WINDOWS_32_BIT)
